Refactor main logic into modules
